Fixed overflow bug where we were fetching a long field as an integer.

+++ src/com/gdrivefs/simplecache/File.java
@@ -1023,8 +1023,8 @@ public class File
 		long currentPosition = start;
 		for(DatabaseRow fragment : fragments)
 		{
-			int startbyte = fragment.getInteger("STARTBYTE");
-			int endbyte = fragment.getInteger("ENDBYTE");
+			long startbyte = fragment.getLong("STARTBYTE");
+			long endbyte = fragment.getLong("ENDBYTE");
 			String chunkMd5 = fragment.getString("CHUNKMD5");
 			java.io.File cachedChunkFile = File.getCacheFile(chunkMd5);
